fix(stacktrace): correct shortFuncName for pointer-typed generic args
The strings.NewReplacer("(", "", "*", "", ")", "") pass introduced by the
previous commit replaced every occurrence of each character, not just the
first one. This is wrong whenever a generic receiver's type argument
itself contains '*', e.g. "(*ServiceLazy[*int]).getInstance": it stripped
the '*' inside the brackets too, producing "ServiceLazy[int].getInstance"
instead of "ServiceLazy[*int].getInstance". Added by a suggested fix on
the PR and confirmed by a new TestShortFuncName case.

Fixed by matching the 2-character "(*" pattern (with a bare "(" fallback)
instead of separate "(" and "*" patterns, so NewReplacer's longest-match
rule only strips the pointer-receiver decoration.

Also moved the tests into the existing stacktrace_test.go (rather than a
separate shortfuncname_test.go) and renamed the benchmark file to
stacktrace_bench_test.go, matching the source file name.

Note: with a 2-byte pattern in the mix, strings.Replacer no longer uses
its single-byte fast path, so this fix is not a net speed win over the
original 3 sequential strings.Replace calls -- it trades a small amount
of speed for fewer allocated bytes. Benchmarked against the pre-PR3
original directly:

goos: darwin / goarch: arm64 / cpu: Apple M3
                 |   before    |            after             |
                 |   sec/op    |   sec/op     vs base          |
ShortFuncName-8   116.2n +- 1%   125.0n +- 1%  +7.53% (p=0.000 n=16)

                 |  before   |           after             |
                 |   B/op    |   B/op     vs base          |
ShortFuncName-8   96.00 +- 0%   88.00 +- 0%  -8.33% (p=0.000 n=16)
                 | allocs/op | allocs/op    vs base        |
ShortFuncName-8   3.000 +- 0%   3.000 +- 0%  ~ (p=1.000 n=16)
