+++ stacktrace/stacktrace_bench_test.go
